build: separated server and app
